# Add more re-associations for associative binary operations