--- a/compiler/mx.compiler/suite.py
+++ b/compiler/mx.compiler/suite.py
@@ -1330,7 +1330,10 @@ suite = {
     "org.graalvm.compiler.phases.common" : {
       "subDir" : "src",
       "sourceDirs" : ["src"],
-      "dependencies" : ["org.graalvm.compiler.phases"],
+      "dependencies" : [
+        "org.graalvm.compiler.loop",
+        "org.graalvm.compiler.phases"
+      ],
       "annotationProcessors" : [
         "GRAAL_PROCESSOR",
       ],
@@ -1420,7 +1423,6 @@ suite = {
       "subDir" : "src",
       "sourceDirs" : ["src"],
       "dependencies" : [
-     "org.graalvm.compiler.loop",
      "org.graalvm.compiler.phases.common",
        ],
       "annotationProcessors" : ["GRAAL_PROCESSOR"],