--- a/compiler/src/org.graalvm.compiler.nodes/src/org/graalvm/compiler/nodes/calc/SubNode.java
+++ b/compiler/src/org.graalvm.compiler.nodes/src/org/graalvm/compiler/nodes/calc/SubNode.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011, 2019, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2011, 2020, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -122,7 +122,7 @@ public class SubNode extends BinaryArithmeticNode<Sub> implements NarrowableArit
                 return forX;
             }
             if (associative && self != null) {
-                ValueNode reassociated = reassociate(self, ValueNode.isConstantPredicate(), forX, forY, view);
+                ValueNode reassociated = reassociateMatchedValues(self, ValueNode.isConstantPredicate(), forX, forY, view);
                 if (reassociated != self) {
                     return reassociated;
                 }
@@ -146,7 +146,7 @@ public class SubNode extends BinaryArithmeticNode<Sub> implements NarrowableArit
                 return NegateNode.create(forY, view);
             }
             if (associative && self != null) {
-                return reassociate(self, ValueNode.isConstantPredicate(), forX, forY, view);
+                return reassociateMatchedValues(self, ValueNode.isConstantPredicate(), forX, forY, view);
             }
         }
         if (forY instanceof NegateNode) {