It has supported the re-association to associate the values which satisfy the specified criterion like `"(a + 1) + 2 => a + (1 + 2)"`.
However, it needs one of the original operands satisfy the criterion. It cannot handle patterns like `"(a + 1) + (b + 2)"`, which could be optimized to `"(a + b) + (1 + 2)"`.

This patch adds another kind of re-association to expose the matched values as the operand. For example with the constantness criterion, `"(a + 1) + b"` is re-associated to `"(a + b) + 1"`. Used together with the existed re-association, it could create more opportunities for other optimizations like constant-fold and loop-invariant elimination.

Change-Id: If32b49a14dd08ac1ab903878da0f459f452da221