@davleopo 2020-07-13
Hi @XiaohongGong , thanks for your contribution.

I have some general questions: do you have any particular benchmarks/workloads in mind that may profit from the additional re-association here?
And with that, did you measure the compile time impact of the transformations?

Could you elaborate a bit more on the use cases you had in mind, looking at the failing test cases it seems the re-association should
not always be done as it interferes with Graal's counted loop detection with is essential for many optimizations.

There are significant tests failing, the CountedLoopTests.
Your additional re-association seems to trigger for loop induction variables effectively destroying the code patterns we except for induction variable detection with is necessary for loops to be detected as counted.