@XiaohongGong 2020-07-14
> Could you elaborate a bit more on the use cases you had in mind, looking at the failing test cases it seems the re-association should
> not always be done as it interferes with Graal's counted loop detection with is essential for many optimizations.
> 
> There are significant tests failing, the CountedLoopTests.
> Your additional re-association seems to trigger for loop induction variables effectively destroying the code patterns we except for induction variable detection with is necessary for loops to be detected as counted.

Hi @davleopo  Thanks for looking at this PR.  The main reason for me to create this re-association is to make more opportunities  for  the loop "ReassociateInvariantPhase" which tries to hoist the loop-invariant expressions. 

And I have noticed the failing case of `"CountedLoopTest"` as you mentioned above, and the possible influence to the counted loop optimization.  I have already created a fixing for it, that is just skipping the re-association for patterns like `i = i + var` and `i = i - var`.  The new patch will be updated soon. Thanks again!