@XiaohongGong 2020-07-14
> I have some general questions: do you have any particular benchmarks/workloads in mind that may profit from the additional re-association here?
> And with that, did you measure the compile time impact of the transformations?

I will have a testing for the performance and compilation impact. Thanks again!