@XiaohongGong 2020-07-14
Hi @davleopo , here is the compilation and performance test results.

**Compilation**
I got the compilation time by respectively running bootstrap with `"-XX:+BootstrapJVMCI"` and the `CTW` with `mx gate -o -t 'CTW:hosted'`.  And the final value is the average of the first 2000 compilation results.

```
                   without this patch       with this patch      delta
Bootstrap              423.505ms               429.539ms         -0.014 
CTW                    108.001ms               105.412ms         +0.025
```
So I think the compilation impact is tiny enough to be ignore.

**Performance**
Here is a jmh benchmark:
```
@BenchmarkMode(Mode.AverageTime)
@OutputTimeUnit(TimeUnit.NANOSECONDS)
@State(Scope.Benchmark)
@Fork(value = 3)
@Measurement(iterations = 5, time = 1000, timeUnit = TimeUnit.MILLISECONDS)
@Warmup(iterations = 5, time = 1000, timeUnit = TimeUnit.MILLISECONDS)
public class LoopInvariant {

    @Param("1024")
    public int length;

    private int[] ia;
    private long[] la;

    private Random random = new Random(0);
    private int startIndex = 4;
    private int field1 = 4;
    private int field2 = 5;
    private long fieldl1 = 4L;
    private long fieldl2 = 5L;

    /**
     * Initialize. New array objects and set initial values.
     */
    @Setup(Level.Trial)
    public void init() {
        ia = new int[length];
        la = new long[length];
        for (int i = 0; i < length; i++) {
            ia[i] = random.nextInt();
            la[i] = random.nextLong();
        }
    }

    @Benchmark
    public void loopInvariantInt() {
        for (int i = startIndex; i < length; i++) {
            int m = i * field1;
            int n = (i - 1) * field2;
            ia[i] = m * n;
        }
    }

    @Benchmark
    public void loopInvariantLong() {
        for (int i = startIndex; i < length; i++) {
            long m = i * fieldl1;
            long n = (i - 1) * fieldl2;
            la[i] = m * n;
        }
    }
}
```

And the results on x86_64:
```
before 
  Benchmark                        (length)  Mode  Cnt     Score     Error  Units
  LoopInvariant.loopInvariantInt       1024  avgt   15  1503.548 ± 114.386  ns/op
  LoopInvariant.loopInvariantLong      1024  avgt   15  1479.330 ±  17.618  ns/op
after  
  Benchmark                        (length)  Mode  Cnt     Score     Error  Units
  LoopInvariant.loopInvariantInt       1024  avgt   15  1127.063 ± 102.551  ns/op
  LoopInvariant.loopInvariantLong      1024  avgt   15  1291.498 ±  44.286  ns/op
```

Results on AArch64:
```
before
  Benchmark                        (length)  Mode  Cnt     Score   Error  Units
  LoopInvariant.loopInvariantInt       1024  avgt   15  1568.473 ± 0.235  ns/op
  LoopInvariant.loopInvariantLong      1024  avgt   15  1536.475 ± 0.548  ns/op

after   
  Benchmark                        (length)  Mode  Cnt     Score   Error  Units
  LoopInvariant.loopInvariantInt       1024  avgt   15  1182.658 ± 2.619  ns/op
  LoopInvariant.loopInvariantLong      1024  avgt   15  1171.550 ± 0.282  ns/op
```