@XiaohongGong 2020-07-21
> While re-association inside loops can gain performance, as you outlined, I am not sure about the involved compile time impact
> doing it in regular canonicalizations, that is why we limited it so far.
> Doing it multi level in canonical methods unconditionally seems a potential compile time problem.
> Can you run with -Dgraal.Time (e.g. a bootstrap, ctw and some dacapo benchmarks) so we see what the phase local time there is both to canonicalization and re-association phase.

Hi @davleopo sure! I'v got the compile time by running the bootstrap, ctw and dacapo benchmarks. 
Here is the tests command the results:

Bootstrap: 
```
mx vm -server -XX:+UnlockExperimentalVMOptions -XX:+EnableJVMCI  -Dgraal.CompilationWatchDogStartDelay=60.0D -esa -Dgraal.DetailedAsserts=true -Dgraal.VerifyGraalGraphs=true -Dgraal.VerifyGraalGraphEdges=true -Dgraal.VerifyGraalPhasesSize=true -Dgraal.VerifyPhases=true -Dgraal.CompilationFailureAction=ExitVM -Dgraal.Time=FrontEnd:4 -XX:+UseJVMCICompiler -XX:-TieredCompilation -XX:+BootstrapJVMCI -version
```

```
FrontEnd_Accm                   606756.1 ms
CanonicalizerPhase_Accm         100658.9 ms
ReassociateInvariantPhase_Accm  5737.3 ms
```              
CTW:
```
mx gate --extra-vm-argument=-Dgraal.Time=FrontEnd:4 -o -t 'CTW:hosted'
```
```
FrontEnd_Accm                   1648485.7 ms
CanonicalizerPhase_Accm         312858.3 ms
ReassociateInvariantPhase_Accm  17276.4 ms

```

Dacapo:
```
mx benchmark dacapo:* -- -Dgraal.Time=FrontEnd:4
```
```
                                lusearch     jython     h2      pmd     avrora   luindex   fop     xalan    sunflow  unit

FrontEnd_Accm                   1648485.7   298992.0  48478.4  66268.3  20574.2  24677.6  47170.0  55162.3  23199.4  ms
CanonicalizerPhase_Accm         312858.3    38963.8   8566.1   11983.6  3289.0   4099.0   8450.9   9830.8   4019.6   ms
ReassociateInvariantPhase_Accm  17276.4     3122.9    506.3    1098.0   193.9    421.5    441.9    415.3    279.8    ms
```