@XiaohongGong 2020-08-17
Hi @davleopo , I'v updated the patch to the new one. Here is the main changes compared to before:
1. Rename the original "ReassociateInvariantPhase" to "ReassociationPhase".  It not only contains the loop invariant re-associations, but also the constant re-associations (only the constant push down part "(a * 2) * b" => "(a * b) * 2").   To avoid regressions to loop invariant, the constant part is applied after loop invariant re-association, and only applied to expressions outside of the loop.
2. For "ReassociationPhase", to avoid collecting the "loopDatas" each time the nodes in loops are changed (it may increase the compile time), the new added loop expressions are saved and appended to the original loop nodes list.
3. Canonicalization for "multiply constant" ("m * 2"  => "m << 1") is deferred after the "ReassociationPhase", to make sure expressions like "(m * 4) * (n * 2)" could be reassociated to "(m * n) * 8", and then to "(m * n) << 3" finally.

I'v tested our internal benchmarks and didn't find the significant performance regressions.

The new compile time compared with master is:
```
Bootstrap:
                                              Before        After       unit
FrontEnd_Accm                                 490614.7     493758.8      ms
PhaseTime_CanonicalizerPhase_Accm             36398.6      33296.6       ms
PhaseTime_CanonicalizerPhase_Instance_Accm    47496.3      45102.7       ms
PhaseTime_ReassociateInvariantPhase_Accm      2751.7       5411.4        ms

CTW:
                                              Before        After       unit
FrontEnd_Accm                                 1480048.7    1478955.4     ms
PhaseTime_CanonicalizerPhase_Accm             112048.9     110754.5      ms
PhaseTime_CanonicalizerPhase_Instance_Accm    183582.1     182383.9      ms
PhaseTime_ReassociateInvariantPhase_Accm       9641.1      18727.1       ms

Dacapo:

Without this patch:
                                  lusearch     jython     h2      pmd     avrora   luindex   fop     xalan    sunflow  unit

FrontEnd_Accm                     38414.6     344545.8  51026.5  94873.1  19802.0  29773.0  62950.5  68017.5  42132.7   ms
CanonicalizerPhase_Accm           3125.5      24713.7   4630.2   8525.2   1657.3   2661.0   5714.3   5931.6   3975.1    ms
CanonicalizerPhase_Instance_Accm  3848.3      29215.3   5379.9   9438.6   1944.8   2763.8   6663.2   7584.1   4698.5    ms
ReassociateInvariantPhase_Accm    171.3       2502.4    295.3    662.9    75.6     304.9    410.4    333.1    161.2     ms

With this patch:
                                  lusearch     jython     h2      pmd     avrora   luindex   fop     xalan    sunflow  unit

FrontEnd_Accm                     34402.9    293843.7   46611.3  67609.8   18022.4  23882.6  42556.1  45907.5  27409.5  ms
CanonicalizerPhase_Accm           2722.3     18357.7    4347.9   6339.2    1504.2   2078.8   3853.1   3957.8   2472.2   ms
CanonicalizerPhase_Instance_Accm  3590.4     22290.1    4753.0   6645.4    1773.8   2330.4   4478.3   4999.7   2889.7   ms
ReassociationPhase_Accm           312.6      3367.5     421.6    799.5     151.7    391.5    404.8    391.7    215.3    ms
```

The time for the single phase `ReassociationPhase` increased, however  the whole compile time can be reduced in a way. So I think the compile time is ok. 

Regarding to your feedback! Thanks!